+++ hdsr_fewspy/constants/request_settings.py
@@ -11,15 +11,15 @@ class RequestSettings:
     default_request_period: pd.Timedelta = None
     min_time_between_requests: pd.Timedelta = None
     max_response_time: pd.Timedelta = None  # Warn if response time is above and adapt next request
-    max_request_size_kb: int = None  # Warn if request size [kb] is above and adapt next request
+    updated_request_period: pd.Timedelta = None
 
 
-default_request_settings = RequestSettings(
-    max_request_nr_timestamps=100000,
-    min_request_nr_timestamps=10000,
-    max_request_period=pd.Timedelta(weeks=52 * 2),
-    default_request_period=pd.Timedelta(weeks=5),
-    min_time_between_requests=pd.Timedelta(seconds=1),
-    max_response_time=pd.Timedelta(seconds=20),
-    max_request_size_kb=3000,
-)
+def get_default_request_settings():
+    return RequestSettings(
+        max_request_nr_timestamps=100000,
+        min_request_nr_timestamps=10000,
+        max_request_period=pd.Timedelta(weeks=52 * 2),
+        default_request_period=pd.Timedelta(weeks=8),
+        min_time_between_requests=pd.Timedelta(seconds=1),
+        max_response_time=pd.Timedelta(seconds=20),
+    )